% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bios.R
\name{get_bio}
\alias{get_bio}
\title{Get speller biography}
\usage{
get_bio(season, id)
}
\arguments{
\item{season}{Season of competition}

\item{id}{ID number of the speller}
}
\description{
Get speller biography
}

