% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bios.R
\name{get_bios}
\alias{get_bios}
\title{Get all biographies of spellers in multiple seasons}
\usage{
get_bios(seasons)
}
\arguments{
\item{seasons}{Vector of seasons}
}
\description{
Get all biographies of spellers in multiple seasons
}

