% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dict.R
\name{get_dict}
\alias{get_dict}
\title{Get dictionary information for word}
\usage{
get_dict(word)
}
\arguments{
\item{word}{Word to lookup in dictionary}
}
\description{
Get dictionary information for word
}

