% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_round_results}
\alias{get_round_results}
\title{Get round results}
\usage{
get_round_results(round, season)
}
\arguments{
\item{round}{Round number}

\item{season}{Season (e.g. 2012)}
}
\description{
Get all results from a single round in the spelling bee
}

