% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bios.R
\name{get_season_bios}
\alias{get_season_bios}
\title{Get biographies of all spellers in a season}
\usage{
get_season_bios(season)
}
\arguments{
\item{season}{Season of competition}
}
\description{
Get biographies of all spellers in a season
}

