% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_season_rounds}
\alias{get_season_rounds}
\title{Get season rounds}
\usage{
get_season_rounds(season)
}
\arguments{
\item{season}{Season (e.g. 2012)}
}
\description{
Get all rounds in a single competition
}

