% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dict.R
\name{get_words}
\alias{get_words}
\title{Get list of competition words}
\usage{
get_words(results)
}
\arguments{
\item{results}{Results from previous competitions, generated by [get_seasons]}
}
\description{
Get list of competition words
}

